merge-ort: add data structures for allowable trivial directory resolves
For files, if two of the three sides match, we can take one of the sides
as the resolution without any further thought.  We would really like to
do the same thing with directories, because being able to do a trivial
directory resolve means we don't have to recurse into the directory,
potentially saving us a huge amount of time in both collect_merge_info()
and process_entries().  However, the possibility of renames means it is
not safe to always resolve such directories.

Sometimes, as humans, we know that sometimes renames don't affect
certain directories.  Unfortunately, rename detection comes after
collect_merge_info() rather than before.  But if there were a way to
learn that rename detection wouldn't need anything in certain
directories, then we could do trivial resolves for those directories.
Interestingly, we know for certain there is one case where this is
possible; merge-recrusive already had a trivial directory resolution
optimization in one very specific case: when the toplevel directory
matches the merge base on one of the two sides of history.

Analyzing why we can do the trivial directory resolve when the toplevel
directory matches between the merge base and one of the two sides, in
conjunction with our knowledge about the optimizations allowing us to
skip-irrelevant-rename-detection allows us to uncover conditions for
why it is okay to do that optimization for the toplevel directory...and
also provides conditions for when we can do the same for subdirectories.

Add some data structures that will allow us to track the necessary
conditions, with some lengthy comments explaining their purpose and use;
future commits will add usage of these data structures as well as
deferred recursion into such directories.

Signed-off-by: Elijah Newren <newren@gmail.com>